mix loaddb (mscmp_syst_db v0.1.0)

Loads the identified Datastore "type" into a testing/development oriented database server.

Options

Most options come with reasonable default values for most development/testing scenarios. Specific circumstances may require overriding the default values by explicitly providing the options documented below.

--type - the Datastore "type" which to load to the database. This is a required string value and for which there is no default. If the --build option is also set, the value of this option will indicate the type of the Datastore migrations to build.

--db-host - a string indicating the host address of the database server. This can be an IP address or resolvable DNS entry. The default value is 127.0.0.1.

--db-port - an integer indicating the TCP port on which to contact the database server. The default value is the standard PostgreSQL port number 5432.

--db-name - a binary value indicating a name for the database to use. The default database name is "ms_devsupport_database".

--dbadmin-pwd - a binary value for the standard ms_syst_privileged database role account created via the database bootstrapping script. The default value is "musesystems.publicly.known.insecure.devsupport.password".

--ds-name - The Datastore Name used to identify the Datastore in the application. Specifically references parts of the supervisory system which manage the Datastore Contexts. This is an optional string value which defaults to "ms_devsupport_database".

--ds-code - a binary value providing a Datastore level salting value used in different hashing operations. The default value is "musesystems.publicly.known.insecure.devsupport.code"

--db-salt - a binary value providing a Datastore level salting value used in different hashing operations. The default value is "musesystems.publicly.known.insecure.devsupport.salt"

--db-role-prefix - a binary value which is prefixed to the names of the database roles created to back the Datastore Contexts. The default value is "ms_devsupport".

--dbadmin-pool - the number of database connections which will be opened to support DBA or Privileged operations. The default value is 1.

--context-name - a binary value which provides a unique context name for the login Context identified by this function. The default value is "ms_devsupport_context".

--context-pwd - a binary value which is the database password that the login Datastore Context uses to log into the database. The default value is "musesystems.publicly.known.insecure.devsupport.apppassword".

--context-pool - the number of database connections the login Context will establish from the application. The default value is 5.

--desc-prefix - a binary value which is prefixed to the descriptions of the created database contexts and which appear in the database role descriptions. The default value is "Muse Systems DevSupport".

--build - a switch which, if set, indicates that the Mix.Task.Builddb process should be called to build the Datastore migrations prior to migrating the new Datastore. See the documentation for Mix.Tasks.Builddb for more information about that process. By default, Datastore migrations are not built.

--source - only applies if the --build option is set. Establishes the database source code directory from which to build migrations. See the documentation for Mix.Tasks.Builddb for more.

--destination - only applies if the --build option is set. Sets the destination path into which Datastore migrations will be built. See the documentation for Mix.Tasks.Builddb for more.

--clean - a switch which indicates if the Datastore migrations to be built should be deleted and entirely rebuilt from the database source code. This switch is only used if the --build option is also set. See the documentation for Mix.Tasks.Builddb for more.

Functions

drop_database(datastore_options, opts)

@spec drop_database(MscmpSystDb.Types.DatastoreOptions.t(), Keyword.t()) :: :ok

get_datastore_options(opts \\ [])

@spec get_datastore_options(Keyword.t()) :: MscmpSystDb.Types.DatastoreOptions.t()

load_database(datastore_options, datastore_type)

@spec load_database(MscmpSystDb.Types.DatastoreOptions.t(), String.t()) ::
  {:ok, [String.t()]} | {:error, MscmpSystError.t()}

run(args)

@spec run([binary()]) :: :ok

Callback implementation for Mix.Task.run/1.