MscmpSystDb.Datastore (mscmp_syst_db v0.1.0)

Provides basic OTP related features for Datastores.

Functions in this module support running Datastores under supervision and creates a Datastore supervisor for managing Datastore Context worker processes.

Functions

child_spec(datastore_options, opts \\ [])

@spec child_spec(MscmpSystDb.Types.DatastoreOptions.t(), Keyword.t()) ::
  Supervisor.child_spec()

Provides a Datastore child specification for use with supervisors.

Note that this function is not child_spec/1 as expected when a Supervisor attempts to start a process based on the module name alone. This is because there is a required parameter which renders that method of starting a process invalid as there's no default value for the parameter (datastore_options).

Parameters

datastore_options - a required Map of values which describe the Datastore and Datastore Context related connection options. See MscmpSystDb.Types.DatastoreOptions.t/0 for more.

opts - a Keyword list of options. See the "Options" section for details.

Options

:datastore_name (Msutils.Types.Process.name/0 or nil) - Specifies the name for the Datastore Supervisor. If this option is not provided, the datastore_options.datastore_name value will be used as the default name for the Datastore Supervisor. If this value identifies a process registry (e.g. {:via, Registry, {MyApp.Registry, :my_registry}}), this registry will become the default registry for all Datastore Contexts; a valid context_registry value overrides this default.

:context_registry (Msutils.Types.Process.registry/0 or nil) - Specifies the registry to use for registering named Datastore Contexts. Can be :local, :global, a tuple of {module(), term()}, or nil. Commonly, this is a tuple of {Registry, registry_name}.

:strategy (atom/0) - Specifies the restart strategy for the Datastore Supervisor. The default value is :one_for_one.

:restart (atom/0) - Specifies the restart strategy for the Datastore Supervisor. The default value is :transient.

:timeout (timeout/0) - Specifies the timeout value for the Datastore Supervisor. The default value is 60000.

start_link(datastore_options, opts \\ [])

@spec start_link(MscmpSystDb.Types.DatastoreOptions.t(), Keyword.t()) ::
  Supervisor.on_start()

Starts the Datastore Supervisor and its requested child Datastore Contexts.

When this function is called, a new DynamicSupervisor for the datastore is started and any Datastore Contexts which are identified in the datastore_options are also started and places under the Datastore Supervisor.

Parameters

datastore_options - a required Map of values which describe the Datastore and Datastore Context related connection options. See MscmpSystDb.Types.DatastoreOptions.t/0 for more.

opts - See the "Options" section for details.

Options

:datastore_name (Msutils.Types.Process.name/0 or nil) - Specifies the name for the Datastore Supervisor. If this option is not provided, the datastore_options.datastore_name value will be used as the default name for the Datastore Supervisor. If this value identifies a process registry (e.g. {:via, Registry, {MyApp.Registry, :my_registry}}), this registry will become the default registry for all Datastore Contexts; a valid context_registry value overrides this default.

:context_registry (Msutils.Types.Process.registry/0 or nil) - Specifies the registry to use for registering named Datastore Contexts. Can be :local, :global, a tuple of {module(), term()}, or nil. Commonly, this is a tuple of {Registry, registry_name}.

:strategy (atom/0) - Specifies the restart strategy for the Datastore Supervisor. The default value is :one_for_one.

:restart (atom/0) - Specifies the restart strategy for the Datastore Supervisor. The default value is :transient.

:timeout (timeout/0) - Specifies the timeout value for the Datastore Supervisor. The default value is 60000.