MscmpSystDb.Types.BoundsCompareResult (mscmp_syst_db v0.1.0)

The comparison operators for both the lower and upper bounds of a range type.

There are cases where normal comparisons are too coarse-grained to provide a meaningful result when dealing with ranges. In these cases you need the detailed lower/upper comparison results.

Types

t()

@type t() :: %MscmpSystDb.Types.BoundsCompareResult{
  lower_comparison: MscmpSystDb.Types.db_type_comparison_operators(),
  upper_comparison: MscmpSystDb.Types.db_type_comparison_operators()
}

The comparison operators for both the lower and upper bounds of a range type.

See MscmpSystDb.Types.BoundsCompareResult for more.