MscmpSystDb.Types.ContextState (mscmp_syst_db v0.1.0)

A struct for reporting the runtime state of Datastore Contexts.

t()

@type t() :: %MscmpSystDb.Types.ContextState{
  context: MscmpSystDb.Types.context_name(),
  state: MscmpSystDb.Types.context_state_values()
}

A struct for reporting the runtime state of Datastore Contexts.

Attributes

:context - identifies the name of the Datastore Context for which runtime state is being reported.

:state - indicates the current runtime state of the named Datastore Context. See MscmpSystDb.Types.context_state_values/0 for more.