MscmpSystDb (mscmp_syst_db v0.1.0)

MscmpSystDb - Database Connectivity & Migrations

The public API for database interactions.

Query

aggregate(queryable, aggregate, field, opts \\ [])

See MscmpSystDb.Runtime.Datastore.aggregate/4.

all(queryable, opts \\ [])

@spec all(Ecto.Queryable.t(), Keyword.t()) :: [Ecto.Schema.t()]

See MscmpSystDb.Runtime.Datastore.all/2.

delete(struct_or_changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.delete/2.

delete!(struct_or_changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.delete!/2.

delete_all(queryable, opts \\ [])

See MscmpSystDb.Runtime.Datastore.delete_all/2.

exists?(queryable, opts \\ [])

See MscmpSystDb.Runtime.Datastore.exists?/2.

get(queryable, id, opts \\ [])

See MscmpSystDb.Runtime.Datastore.get/3.

get!(queryable, id, opts \\ [])

See MscmpSystDb.Runtime.Datastore.get!/3.

get_by(queryable, clauses, opts \\ [])

See MscmpSystDb.Runtime.Datastore.get_by/3.

get_by!(queryable, clauses, opts \\ [])

See MscmpSystDb.Runtime.Datastore.get_by!/3.

in_transaction?()

See MscmpSystDb.Runtime.Datastore.in_transaction?/0.

insert(struct_or_changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.insert/2.

insert!(struct_or_changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.insert!/2.

insert_all(schema_or_source, entries_or_query, opts \\ [])

See MscmpSystDb.Runtime.Datastore.insert_all/3.

insert_or_update(changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.insert_or_update/2.

insert_or_update!(changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.insert_or_update!/2.

load(module_or_map, data)

See MscmpSystDb.Runtime.Datastore.load/2.

one(queryable, opts \\ [])

See MscmpSystDb.Runtime.Datastore.one/2.

one!(queryable, opts \\ [])

See MscmpSystDb.Runtime.Datastore.one!/2.

preload(structs_or_struct_or_nil, preloads, opts \\ [])

See MscmpSystDb.Runtime.Datastore.preload/3.

prepare_query(operation, query, opts \\ [])

See MscmpSystDb.Runtime.Datastore.prepare_query/3.

query_for_many(query, query_params \\ [], opts \\ [])

@spec query_for_many(iodata(), [term()], Keyword.t()) ::
  {:ok,
   %{
     :rows => nil | [[term()] | binary()],
     :num_rows => non_neg_integer(),
     optional(atom()) => any()
   }}
  | {:error, MscmpSystError.t()}

Executes a database query and returns all rows.

query_for_many!(query, query_params \\ [], opts \\ [])

@spec query_for_many!(iodata(), [term()], Keyword.t()) :: %{
  :rows => nil | [[term()] | binary()],
  :num_rows => non_neg_integer(),
  optional(atom()) => any()
}

Executes a database query and returns all rows. Raises on error.

query_for_none(query, query_params \\ [], opts \\ [])

@spec query_for_none(iodata(), [term()], Keyword.t()) ::
  :ok | {:error, MscmpSystError.t()}

Executes a database query but returns no results.

query_for_none!(query, query_params \\ [], opts \\ [])

@spec query_for_none!(iodata(), [term()], Keyword.t()) :: :ok

Executes a database query but returns no results. Raises on error.

query_for_one(query, query_params \\ [], opts \\ [])

@spec query_for_one(iodata(), [term()], Keyword.t()) ::
  {:ok, [any()]} | {:error, MscmpSystError.t()}

Executes a database query and returns a single row.

query_for_one!(query, query_params \\ [], opts \\ [])

@spec query_for_one!(iodata(), [term()], Keyword.t()) :: [any()]

Executes a database query and returns a single row. Raises on error.

query_for_value(query, query_params \\ [], opts \\ [])

@spec query_for_value(iodata(), [term()], Keyword.t()) ::
  {:ok, any()} | {:error, MscmpSystError.t()}

Executes a database query returning a single value.

query_for_value!(query, query_params \\ [], opts \\ [])

@spec query_for_value!(iodata(), [term()], Keyword.t()) :: any()

Executes a database query returning a single value. Raises on error.

record_count(queryable, opts)

See MscmpSystDb.Runtime.Datastore.record_count/2.

reload(struct_or_structs, opts \\ [])

See MscmpSystDb.Runtime.Datastore.reload/2.

reload!(struct_or_structs, opts \\ [])

See MscmpSystDb.Runtime.Datastore.reload!/2.

rollback(value)

See MscmpSystDb.Runtime.Datastore.rollback/1.

stream(queryable, opts \\ [])

See MscmpSystDb.Runtime.Datastore.stream/2.

transaction(job, opts \\ [])

@spec transaction(
  (... -> any()) | Ecto.Multi.t(),
  keyword()
) :: {:error, MscmpSystError.t()} | {:ok, any()}

See MscmpSystDb.Runtime.Datastore.ecto_transaction/2.

update(changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.update/2.

update!(changeset, opts \\ [])

See MscmpSystDb.Runtime.Datastore.update!/2.

update_all(queryable, updates, opts \\ [])

See MscmpSystDb.Runtime.Datastore.update_all/3.

Datastore Management

create_datastore(datastore_options, opts \\ [])

@spec create_datastore(MscmpSystDb.Types.datastore_options(), Keyword.t()) ::
  {:ok, MscmpSystDb.Types.database_state_values(),
   [MscmpSystDb.Types.context_state()]}
  | {:error, MscmpSystError.t()}

Creates a new Datastore along with its contexts.

The creation of a new Datastore includes creating new database to back the Datastore and database roles representing each of the Datastore contexts.

create_datastore_contexts(datastore_options, datastore_contexts, opts \\ [])

@spec create_datastore_contexts(
  MscmpSystDb.Types.datastore_options(),
  [MscmpSystDb.Types.datastore_context(), ...],
  Keyword.t()
) ::
  {:ok, [MscmpSystDb.Types.context_state(), ...]} | {:error, MscmpSystError.t()}

Creates database roles to back all requested Datastore contexts.

Usually Datastore contexts are created in the create_datastore/1 call, but over the course of time it is expected that applications may define new contexts as needs change. This function allows applications to add new contexts to existing Datastores.

drop_datastore(datastore_options, opts \\ [])

@spec drop_datastore(MscmpSystDb.Types.datastore_options(), Keyword.t()) ::
  :ok | {:error, MscmpSystError.t()}

Drops a Datastore along with its contexts.

Dropping a Datastore will drop the database backing the Datastore from the database server as well as all of the database roles associated defined by the provided database options.

Prior to dropping the Datastore, all active connections to the Datastore should be terminated or the function call could fail.

Note that this is am irreversible, destructive action. Any successful call will result in data loss.

drop_datastore_contexts(datastore_options, datastore_contexts, opts \\ [])

@spec drop_datastore_contexts(
  MscmpSystDb.Types.datastore_options(),
  [MscmpSystDb.Types.datastore_context(), ...],
  Keyword.t()
) :: :ok | {:error, MscmpSystError.t()}

Drops the requested Datastore contexts.

This function will drop the database roles from the database server that correspond to the requested Datastore contexts. You should be sure that the requested Datastore contexts do not have active database connections when calling this function as active connections are likely to result in an error condition.

get_datastore_context_states(datastore_contexts, opts \\ [])

@spec get_datastore_context_states(MscmpSystDb.Types.datastore_options(), Keyword.t()) ::
  {:ok, [MscmpSystDb.Types.context_state(), ...]} | {:error, MscmpSystError.t()}

Returns the state of the requested contexts.

This function will check for each given context that: it exist, whether or not database connections may be started for it, and whether or not database connections have been started.

Note that only startable contexts are included in this list. If the context is not startable or has id: nil, the context will be excluded from the results of this function.

get_datastore_state(datastore_options, opts \\ [])

@spec get_datastore_state(MscmpSystDb.Types.datastore_options(), Keyword.t()) ::
  {:ok, MscmpSystDb.Types.database_state_values(),
   [MscmpSystDb.Types.context_state()]}
  | {:error, MscmpSystError.t()}

Returns the state of the database and database roles which back the Datastore and contexts, respectively, of the provided Datastore options definition.

Datastore Migrations

get_datastore_version(datastore_options, opts \\ [])

@spec get_datastore_version(MscmpSystDb.Types.datastore_options(), Keyword.t()) ::
  {:ok, String.t()} | {:error, MscmpSystError.t()}

Returns the most recently installed database migration version number.

upgrade_datastore(datastore_options, datastore_type, migration_bindings, opts \\ [])

@spec upgrade_datastore(
  MscmpSystDb.Types.datastore_options(),
  String.t(),
  Keyword.t(),
  Keyword.t()
) :: {:ok, [String.t()]} | {:error, MscmpSystError.t()}

Updates a Datastore to the most current version of the given type of Datastore.

If a Datastore is already up-to-date, this function is basically a "no-op" that returns the current version. Otherwise, database migrations for the Datastore type are applied until the Datastore is fully upgraded to the most recent schema version.

Runtime

current_datastore_context()

@spec current_datastore_context() :: atom() | pid()

See MscmpSystDb.Runtime.Datastore.current_datastore_context/0.

put_datastore_context(context)

@spec put_datastore_context(pid() | Ecto.Repo.t() | Ecto.Adapter.adapter_meta()) ::
  atom() | pid()

See MscmpSystDb.Runtime.Datastore.put_datastore_context/1.

start_datastore(datastore_options, supervisor_name \\ nil)

@spec start_datastore(
  MscmpSystDb.Types.datastore_options(),
  Supervisor.supervisor() | nil
) ::
  {:ok, :all_started | :some_started,
   [MscmpSystDb.Types.context_state_values()]}
  | {:error, MscmpSystError.t()}

Starts database connections for all of login contexts in the Datastore options.

start_datastore_context(datastore_options, context)

@spec start_datastore_context(
  MscmpSystDb.Types.datastore_options(),
  atom() | MscmpSystDb.Types.datastore_context()
) :: {:ok, pid()} | {:error, MscmpSystError.t()}

Starts a database connection for the specific Datastore context provided.

stop_datastore(datastore_options_or_contexts, db_shutdown_timeout \\ 60000)

@spec stop_datastore(
  MscmpSystDb.Types.datastore_options()
  | [MscmpSystDb.Types.datastore_context()]
  | [%{context_name: MscmpSystDb.Types.context_name()}],
  non_neg_integer()
) :: :ok | {:error, MscmpSystError.t()}

Disconnects the database connections for all of the login Datastore option contexts.

stop_datastore_context(context, db_shutdown_timeout \\ 60000)

@spec stop_datastore_context(
  pid() | atom() | MscmpSystDb.Types.datastore_context(),
  non_neg_integer()
) ::
  :ok

Disconnects the database connection for the specific Datastore context provided.