MscmpSystDb.Types.DatastoreOptions (mscmp_syst_db v0.1.0)

A struct defining the connection parameters to use to connect to a Datastore backing database.

t()

@type t() :: %MscmpSystDb.Types.DatastoreOptions{
  contexts: [MscmpSystDb.Types.DatastoreContext.t()] | [],
  database_name: String.t(),
  datastore_code: String.t() | nil,
  datastore_name: atom() | nil,
  db_server: MscmpSystDb.Types.DbServer.t()
}

A struct defining the connection parameters to use to connect to a Datastore backing database.

Attributes

:database_name - The name of the database in the database server to which the connection will be made. Often times this value will be the same as the String.t() form of the :datastore_name value.

:datastore_code - Defines a Datastore specific salting value for use in certain security and cryptographic related functions.

:datastore_name - A name for use by the application to identify a given Datastore. This value will often time be the same as the :database_name value, except as an atom() rather than a String.t().

:contexts - A list of available Datastore Context values defining which contexts are available for this Datastore. See MscmpSystDb.Types.DatastoreContext.t/0 for more information concerning the values of each entry in the list.

:db_server - The database server information for the server where the Datastore database resides. See 't:MscmpSystDb.Types.DbServer.t/0' for additional details.